bot/requirements.txt
python-telegram-bot==20.3
psycopg2-binary==2.9.6